mcp-servers/entra-graph-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "entra-graph-mcp"
version = "0.1.0"
description = "MCP server for Entra identity management via Microsoft Graph"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "azure-identity>=1.17.0",
    "msgraph-sdk>=1.5.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 120
